OCR_Extraction/requirements.txt
python-docx
opencv-python
pytesseract